feat: student overview should have a title
This is my first test in my test driven design and I have to know the following:
  - my tests have to help me in the design of my software
  - I want to create the simplest test that fails, then add the simplest impl
  - the test is an hypotesis of what my component should do

So, I'm blocked with what to do... I wanted to show straight away the list of students, but then I realized that there are some things that are needed that are much simpler.
I want the user to know that he is placed in an overview of students. I'll need some html block to display this.

My first test is saying 'my overview component should show somewhere a title'
  - Does not matter how, but I know that it will be a title!

footer notes:
  - I'm creating my own requirements, so adding or not a title is up to me.
    - In te real life, we'll have a ticket and most of the time the PO will not care about these details until it's already on production
    - By having this reflext of writing test, we will think twice before stating and ask the PO for more info about the ticket
  - UIs change a lot that's why I do not care much if it's a div or a h2 or whatever tag
    - We do know that there have to be something to hold the a title
    - if we have to change it, it will be because the requirement had changed so it will be prefectly fine to remove this test