feat: should display students
I want to show the students if there are any.

This could be done in a more complex way with ngFor and showing some student info. But, we have to go step by step. This will ensure that we are thinking well all changes.

For the moment, my students are only a string, I havent bothered yet to define what will have an student.