feat: should display every student name
Here I started defining what's an student, this commit says that the student have a name and it should be displayed in the overview view.
  - In real life this constraints should be defined in your ticket, so you could add all student properties in a single step. In my case, I'm deciding at each step what to do.

The data structure is really important wen designing an UI. In most cases, we designed a view where we need a lot of data so we end up having to gather al the data in a single endpoint or calling a lot of endpoints and joining the data in the View component.
  - Imagine I add to this student overview it's grades and the grades are another data structure. Then, my view will have to handle an extra responsibility and I'll be impacting the backbend.
  - What to do then if I have to add grades? I will create a separate view for that, a pop-up or sidebar or other view component.
  - For this example, having the grades with the students does not seem like a big deal and it isn't. The big deal is when we add over the time more data structures to the same view, so it's better to not start doing it.

In my opinion, this student's overview should only show things about the simple properties of an students like name, surname, birth date, enroll date.
  - We could also show some interesting things only related to the students simple properties like:
    - the total number of students
    - the percentage of students that enrolled last month or last week
    - the average age of a student
    - etc