refactor: use fast-check to create fake students
It's annoying to mock data my hand. I've decided to use the wonderful fast-check library to fake the data.

To fake an student we need to create a fast-check record arbitrary object with all the student properties. Each property will be a type arbitrary.
Then, we generate the fake data using the fast-check sample function.

For this case, I've created a fakeStudents function that will an interface to use fast-check. In case, fast-check get obsolete or I need to change it for a better option this will make my life easier.

I placed the Arbitrary and related functions in the student.model.ts file because it's directly related to the Student model.