feat: add loading property
I want to add indicator when loading students.
There are several ways to add this loading, on my case I choose to create a component Props interface to pass all the properties
  - This has optimization advantages when combined with the OnPush change detection
  - The loading will change when the data will change. They are interdependent. Without the Props and the OnPush we will be triggering twice the change detection

I used TDD to do this but I condense all the steps into a single commit. These are the steps.
  - Add an 'props' property to the WraperComponent class with the structure I want and initialize it with default values
  - Use the property in the student-overview tag with. All test will fail
  - Adapt the tests to pass the students and the loading throw the WrapperComponent
  - Add the Props interface in the Overview class
  - Add a new Input prop 'props' of type Prop initialized with default values.
  - Adapt the view to use the new props. All test pass now!
  - Delete the student props