feat: use loading property
Another iterative work with TDD. My componet is able to get some students and change depending on the properties.
What I want to appear is in unit tests, so we check that the component follows the ticket requirements (at least for the content).

This component is a visual component. It does not have any logic, so we are finish with him for now :)