feat: should fetch students and pass it to view component
There are multiple ways to accomplish what this commit is doing, fetch data and forward it to another component.
I've deiced that a service will fetch the data. But it gets complicated because it's remote data because it's an observable.
To avoid boilerplate with the subscription I used the async pipe. This will ensure the subscription is closed correctly.

Test:
```
  it('should fetch students and pass them to the child component', async () => {
    const students = fakeStudents();
    mockStudentService(of(students));
    fixture.detectChanges();

    await fixture.whenStable();

    const studentsOverviewUi: OverviewUiComponent = fixture.debugElement.query(
      By.directive(OverviewUiComponent)
    ).componentInstance;
    expect(studentsOverviewUi.props.students).toEqual(students);
  });

  const mockStudentService =
    (getStudentsMockResult: Observable<Student[]>) => {
      jest.spyOn(studentService
        , 'getStudents'
      ).mockReturnValue(getStudentsMockResult);

      fixture = TestBed.createComponent(OverviewComponent);
      component = fixture.componentInstance;
    }
```

This test maybe looks a bit weird. I'm using jest spyOn to mock the service responses. Also, I'd to use the whenStable method because I'm using an async pipe, so we makes sure that the DOM is correclty rendered.
Because jest has issues with JS when mocking [link](https://stackoverflow.com/questions/56594114/jest-angular-change-value-of-spyon-mockreturnvalue), I had to add this mockStudentService method so it's correcly mocked.
  - There may be other solutions, please contact me if you find anything better :)

Footer notes:
  - Maybe the coding style is weird, I'm trying to use ElmLang style.
  - I've extracted the function that maps the students to the ui properties because when using rxjs we pipe a lot of things and it could be horrible in the future.
    - This is a preference
  - I had to provide a dummy Service impl so the compiler builds the project