feat: should show inital content when fetching the students
We have to make sure to show something to the user when the request is called. On this case we set the loading to true.

Test:
```
  it('should pass initial data to the child component', async () => {
    mockStudentService(new Observable());
    fixture.detectChanges();

    await fixture.whenStable();

    const studentsOverviewUi: OverviewUiComponent = fixture.debugElement.query(
      By.directive(OverviewUiComponent)
    ).componentInstance;
    expect(studentsOverviewUi.props.students).toEqual([]);
    expect(studentsOverviewUi.props.loading).toBe(true);
  });
```

Here I'm passing an empty observable to simulate an on going request. This helps me to verify that we are starting with the correct values, loading true and empty students.