feat(student-list): should contain student's firstname
We have to check if the students properties from the ticket are being shown.
The easier test is to check for one of the properties, the firstname.

At this stage we do not care about how it would look because the ui is intercheangable and we could deal with that later.

First, I created the test. Check if the student.firstname is rendered by angular.
  - Somehow I have to idenfity a student from other students in the html for testing. I need an Id or class specific for each student
Then, I got my first error. The property firstname is not defined. So, I created an interface Student to hold student data.
I execute the test and ... It failed. The firstname is not being rendered.
Let's improve the code! Instead of any I used the Student interface and I used at the UI component, very simply, the student.firstname.
  - Instead of a for loop I could also had used student[0].firstname, but it adds an extra step for the same complexity.
Run the test again and bum!! Sucess :D

Also, because Student is an entity I added an Id to the interface. This will also help me with the ngFor.
- This is also really hepfull for testing, because we can add specific ids for searching on test.