feat(students): add student birthDate, profilepicture and enrolmentDate properties
This hole commit was done with the TDD cycle, test failed, increment, test succeded, for each of the properties one by one.

I combined multiple steps in the same commit because they were really similar. We could also commit each property in a commit, but we could be flexible. The important thing is to add short and simple increments to rollback if needed.

The most special thing here could be the dates and the profile picture.
- With the dates we need to know which format to use, as specified in the ticket it has to be a Date String, like Wed Jun 14 2017
- The profile picture is used in a img src attribute. We need to do some extra work to get it.
  - As we used the student's id as a tag identifier, I dedided that is enough to search the for the img tag inside the student html element.
  - We could change this if we get some complications in the future