refactor(students): improve tests
The refactor is a wider step in TDD. Sometimes is done after a tdd cycle, others after a bunch of cycles.
The idea is to make the bigger picture pretter/faster/better without affecting the visual behaviour.

At this commit I used a better structure for the tests scoping them in 2 parts, one for the No student banner and other for the student properties. I also took advantage of the scope to group the initializaion of the students, so the tests run faster.
I also replaced some double quotes for simple ones to make the quote usage consistent.

When refactoring, we don't need to refactor everything. Just a tiny improvement is enough.