feat(student): replace non styled student list ui with bootstrap-ui
Here we are integrating our new style and all tests are green!!

We first made sure that the data constraint written in the feature ticket are correctly done. To check for this we created tests at the smart component level.

After focusing in this data requirement we did a really simple ui component, but we needed some style. This separation gives us the oportunity to focus on the style and even better, we could parallelize work between 2 people.

We could debate if we should add testing for ui components... In component oriented programing we should completly replace a component instead of adding more things, so if we check for a button to be blue, changing that component shouldn't be that painfull becaus of the tests.
For me everithing is context. Do I need style checks for my ui component ? Nop, *because* I'm at the beginning of the project, this code is for an article so the extra work does not worth it for me now.