feat(student): add student service to get all students
We added to the student domain the components and models folder. This commit includes as well a services folder that will hold all the services that will fetch/submit data.

My context is that the project is just starting, there isn't any backend or somewhere to request the data, so I simulated this with using generated data.