feat(student): use StudentService in StudentListComponent
This change changes the way we get the students. Before the easier way was to use a BehaviourSubject and pass the students.
The tests I did weren't the best... Maybe you already noticed it. I created the component and changed the state (next observable value) for each test. This brokes test isolation but it was simpler.

When using an observable, things get more interesting. I have to mock the StudentService.getAll method return value. This value is an observable instead of a subject, so it's difficult to pass the next value. So, my simpler and nasty approach couldn't work.

To fix this I created a function createTestObjects that will recreate and setup at each test execution the fixture, component and the studentService.getAll return value.
The tests are clean even if this costs a bit more of execution time.