feat(student): add loading indicator
This is part of the feature ticket. We want to be able to tell the user when the data is being requested so he knows something is happening.
I started writing the test, were I find the loading successfuly then do something to simulate the arrival of student's data and finally find that the loading indicator is not there.
The simulation of the incoming data is the difficult part. For this, I adapted the createTestObject function to get an observable of students instead of the students. Then, I used a replay subject to have the control of when the student's data arrives.

I tried to use the new angular concepts, for this I used the signals. Yes, I added more than one thing to the commit. It's not such a problem because the change stays simple and related to the loading indicator.