feat(student): replace non styled loading with bootstrap-loading-ui component
Adding an ui component is a good thing because of reusability, and also it makes an interface for style frameworks like bootstrap.

I'm not sure about the way I'm handling the testing of the loading html element. What happens if there are multiple loadings in the component tree? How do I find them when testing?
For now, I'm leaving it like this to try it out.